requirements.txt
alabaster==0.7.16
albumentations==1.4.3
annotated-doc==0.0.4
anyio==4.12.1
app-model==0.2.8
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
babel==2.18.0
blosc2==4.0.0
build==1.4.0
cachey==0.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dask==2026.1.2
dask-image==2025.11.0
debugpy==1.8.20
decorator==5.2.1
deeplabcut==3.0.0rc13
dlclibrary==0.0.11
docstring_parser==0.17.0
docutils==0.17.1
einops==0.8.2
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.0
filterpy==1.4.5
flexcache==0.3
flexparser==0.4
fonttools==4.61.1
freetype-py==2.5.1
fsspec==2025.12.0
h11==0.16.0
HeapDict==1.0.1
hf-xet==1.2.0
hsluv==5.0.4
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.7.1
in-n-out==0.2.1
ipykernel==6.31.0
ipython==8.38.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
locket==1.0.0
magicgui==0.10.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
napari==0.4.18
napari-console==0.1.4
napari-deeplabcut==0.2.1.7
napari-plugin-engine==0.2.1
napari-svg==0.2.1
natsort==8.4.0
ndindex==1.10.1
nest-asyncio==1.6.0
networkx==3.4.2
npe2==0.7.9
numba==0.63.1
numexpr==2.14.1
numpy==1.26.4
numpydoc==1.5.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==26.0
pandas==2.3.3
parso==0.8.6
partd==1.4.2
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
PIMS==0.7
Pint==0.24.4
platformdirs==4.5.1
polars==1.38.1
polars-runtime-32==1.38.1
pooch==1.9.0
prompt_toolkit==3.0.52
psutil==7.2.2
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.0
pycocotools==2.0.11
pyconify==0.2.1
pydantic==1.10.26
pydantic-compat==0.1.2
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.3.2
pyproject_hooks==1.2.0
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
QDarkStyle==3.1
qtconsole==5.7.1
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.10.2
six==1.17.0
slicerator==1.1.0
snowballstemmer==3.0.1
Sphinx==4.5.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
statsmodels==0.14.6
superqt==0.7.8
sympy==1.13.1
tables==3.10.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
tomli==2.4.0
tomli_w==1.2.0
toolz==1.1.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triton==3.1.0
typer==0.22.0
typer-slim==0.22.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
vispy==0.12.2
wcwidth==0.6.0
wrapt==2.1.1
zipp==3.23.0
